requirements.txt
# Required Python packages get listed here, one per line.

runpod~=1.7.9
torch
torchvision
torchaudio
transformers
diffusers
accelerate
safetensors
pillow
boto3
